+++ R/cranDistribution.R
@@ -20,19 +20,17 @@ cranDistribution <- function(date = NULL, all.filters = FALSE,
   cran_log <- cleanLog(cran_log)
   ymd <- rev_fixDate_2012(file.url.date)
   
+  cores <- multiCore(multi.core)
+  if (.Platform$OS.type == "windows" & cores > 1) cores <- 1L
+  
   if (all.filters) {
     ip.filter <- TRUE
     small.filter <- TRUE
   }
   
   if (small.filter) cran_log <- smallFilter(cran_log)
+  if (ip.filter) cran_log <- ipFilter(cran_log, multi.core = cores)
 
-  if (ip.filter) {
-    cores <- multiCore(multi.core)
-    if (.Platform$OS.type == "windows" & cores > 1) cores <- 1L
-    cran_log <- ipFilter(cran_log, multi.core = cores)
-  }
-  
   freqtab <- sort(table(cran_log$package), decreasing = TRUE)
   
   pkg.data <- data.frame(package = names(freqtab), count = c(freqtab), 
@@ -41,18 +39,15 @@ cranDistribution <- function(date = NULL, all.filters = FALSE,
   rnk <- rank(pkg.data$count, ties.method = "min")
   pkg.data$rank <- (max(rnk) + 1) - rnk
   pkg.data$nominal.rank <- seq_len(nrow(pkg.data))
-  pkg.data$unique.packages <- length(freqtab)
 
-  pkg.data$percentile <- vapply(pkg.data$count, function(x) {
+  pkg.data$percentile <- unlist(parallel::mclapply(pkg.data$count, function(x) {
     round(100 * mean(pkg.data$count < x), 1)
-  }, numeric(1L))
+  }, mc.cores = cores))
   
-  out <- list(date = ymd, data = pkg.data)
+  out <- list(date = ymd, unique.packages = length(freqtab), data = pkg.data)
   class(out) <- "cranDistribution"
   out
 }
-
-mcranDistribution <- memoise::memoise(cranDistribution)
 
 #' Plot method for cranDistribution().
 #' @param x An object of class "cranDistribution" created by \code{cranDistribution()}.
@@ -62,8 +57,9 @@ mcranDistribution <- memoise::memoise(cranDistribution)
 #' @export
 
 plot.cranDistribution <- function(x, type = "count", ...) {
-  ttl <- paste("Download Count Distribution @", x$date)
-  xlab <-  "Log10 Count"
+  day <- weekdays(as.Date(x$date), abbreviate = TRUE)
+  ttl <- paste0("CRAN/Posit @ ", x$date, " (", day, ")")
+  xlab <-  "Log10 Download Count"
   if (type == "histogram") {
     graphics::hist(log10(x$data$count), main = ttl, xlab = xlab)
   } else if (type == "count") {
@@ -87,17 +83,25 @@ plot.cranDistribution <- function(x, type = "count", ...) {
       col.axis = "red", col.ticks = "red")
     axis(3, at = log10(max), cex.axis = 0.8, padj = 0.9, labels = max.lab)
   } else stop('type must be "historgram" or "count"', call. = FALSE)
-  title(sub = paste(format(x$data$unique.packages[1], big.mark = ","), 
-    "unique packages"), cex.sub = 0.9)
+  title(sub = paste(
+    format(sum(x$data$count), big.mark = ","), "total downloads;",
+    format(x$unique.packages, big.mark = ","), "unique packages"), 
+    cex.sub = 0.9)
 }
 
 #' Print method for cranDistribution().
 #' @param x object.
+#' @param top.n Numeric or Integer.
 #' @param ... Additional parameters.
 #' @export
 
-print.cranDistribution <- function(x, ...) {
-  print(list(x$date, head(x$data, 20)))
+print.cranDistribution <- function(x, top.n = 20, ...) {
+  pkg.ct <- format(x$unique.packages, big.mark = ",")
+  dwnld.ct <- format(sum(x$data$count), big.mark = ",")
+  print(list(date = x$date, 
+             unique.packages.downloaded = pkg.ct,
+             total.downloads = dwnld.ct,
+             top.n = head(x$data, top.n)))
 }
 
 #' Summary method for cranDistribution().
@@ -109,7 +113,9 @@ print.cranDistribution <- function(x, ...) {
 #' @export
 
 summary.cranDistribution <- function(object, ...) {
-  summary(object$data$count)
+  list(unique.packages.downloaded = object$unique.packages,
+       total.downloads = sum(object$data$count),
+       download.summary = summary(object$data$count))
 }
 
 #' Query download count.
@@ -128,7 +134,7 @@ queryCount <- function(count = 1, date = NULL, all.filters = FALSE,
   ip.filter = FALSE, small.filter = FALSE, memoization = TRUE, 
   multi.core = FALSE) {
 
-  x <- mcranDistribution(date = date, all.filters = all.filters, 
+  x <- cranDistribution(date = date, all.filters = all.filters, 
     ip.filter = ip.filter, small.filter = small.filter, 
     memoization = memoization, multi.core = multi.core)
 
@@ -145,35 +151,50 @@ queryCount <- function(count = 1, date = NULL, all.filters = FALSE,
 
 #' Query package name.
 #'
-#' @param package Character..
+#' @param packages Character..
 #' @param date Character. Date. "yyyy-mm-dd". NULL uses latest available log.
 #' @param all.filters Logical. Master switch for filters.
 #' @param ip.filter Logical.
 #' @param small.filter Logical. TRUE filters out downloads less than 1000 bytes.
 #' @param memoization Logical. Use memoization when downloading logs.
+#' @param check.package Logical. Validate and "spell check" package.
 #' @param multi.core Logical or Numeric. \code{TRUE} uses \code{parallel::detectCores()}. \code{FALSE} uses one, single core. You can also specify the number logical cores. Mac and Unix only.
 #' @return An R data frame.
 #' @export
 
-queryPackage <- function(package = "packageRank", date = NULL, 
+queryPackage <- function(packages = "packageRank", date = NULL, 
   all.filters = FALSE, ip.filter = FALSE, small.filter = FALSE, 
-  memoization = TRUE, multi.core = FALSE) {
+  memoization = TRUE, check.package = TRUE, multi.core = FALSE) {
   
-  x <- mcranDistribution(date = date, all.filters = all.filters, 
-                           ip.filter = ip.filter, small.filter = small.filter, 
-                           memoization = memoization, multi.core = multi.core)
+  if (check.package) packages <- checkPackage(packages)
+  
+  x <- cranDistribution(date = date, all.filters = all.filters,
+    ip.filter = ip.filter, small.filter = small.filter, 
+    memoization = memoization, multi.core = multi.core)
+
+  unobs.pkgs <- !packages %in% x$data$package
+  if (any(unobs.pkgs)) pkg.msg <- paste(packages[unobs.pkgs], collapse = ", ")
+
+  if (all(unobs.pkgs)) {
+    stop("No downloads for ", pkg.msg, " on ", x$date, ".", call. = FALSE)
+  } else if (any(unobs.pkgs)) {
+    message("No downloads for ", pkg.msg, " on ", x$date, ".")
+    packages <- packages[!unobs.pkgs]
+  }
   
   tmp <- x$data
   
-  if (all(!package %in% tmp$package)) {
+  if (all(!packages %in% tmp$package)) {
     stop("Package(s) not observed. Check spelling.", call. = FALSE)
-  } else if (any(!package %in% tmp$package)) {
-    message("No downloads for ", paste(package[!package %in% tmp$package], 
+  } else if (any(!packages %in% tmp$package)) {
+    message("No downloads for ", paste(packages[!packages %in% tmp$package], 
       collapse = ", "), ".")
-    out <- tmp[tmp$package %in% package, ]
-  } else if (all(package %in% tmp$package)) {
-    out <- tmp[tmp$package %in% package, ]
+    out <- tmp[tmp$package %in% packages, ]
+  } else if (all(packages %in% tmp$package)) {
+    out <- tmp[tmp$package %in% packages, ]
   }
+
+  out <- out[match(packages, out$package), ]
   rownames(out) <- NULL
   out
 }
@@ -195,7 +216,7 @@ queryRank <- function(num.rank = 1, rank.ties = FALSE, date = NULL,
   all.filters = FALSE, ip.filter = FALSE, small.filter = FALSE, 
   memoization = TRUE, multi.core = FALSE) {
   
-  x <- mcranDistribution(date = date, all.filters = all.filters, 
+  x <- cranDistribution(date = date, all.filters = all.filters, 
     ip.filter = ip.filter, small.filter = small.filter, 
     memoization = memoization, multi.core = multi.core)
   
@@ -226,7 +247,7 @@ queryPercentile <- function(percentile = 50, lo = NULL, hi = NULL,
   date = NULL, all.filters = FALSE, ip.filter = FALSE, small.filter = FALSE, 
   memoization = TRUE, multi.core = FALSE) {
   
-  x <- mcranDistribution(date = date, all.filters = all.filters, 
+  x <- cranDistribution(date = date, all.filters = all.filters, 
     ip.filter = ip.filter, small.filter = small.filter, 
     memoization = memoization, multi.core = multi.core)